requirements.txt
flask
pandas
ccxt>=4.3.0
python-dotenv
gunicorn
colorama
